+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## unreleased
 
-* Check if a pending browser switch request exists before delivering a browser switch result instead of setting Activity intent to null (fix for #634)
+* Check if a pending browser switch request exists before delivering a browser switch result instead of setting Activity intent to null
+* Fix issue that causes a browser switch to start while the host Activity is finishing
 
 ## 2.3.1